Add GeneratedClass abstraction to codegen logic.
